+++ pkg/core/zone/api.go
@@ -1,4 +1,4 @@
-package records
+package zone
 
 import (
 	"context"
